requirements.txt
Django==1.10.1
django-dashing==0.3.1